requirements.txt
Django>=4.2.5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
Pillow>=9.0
django-cors-headers>=4.0
